relegate legacy media handlers to separate file
