+++ src/api/client/mod.rs
@@ -9,6 +9,7 @@ pub(super) mod directory;
 pub(super) mod filter;
 pub(super) mod keys;
 pub(super) mod media;
+pub(super) mod media_legacy;
 pub(super) mod membership;
 pub(super) mod message;
 pub(super) mod openid;
@@ -46,6 +47,7 @@ pub(super) use directory::*;
 pub(super) use filter::*;
 pub(super) use keys::*;
 pub(super) use media::*;
+pub(super) use media_legacy::*;
 pub(super) use membership::*;
 pub use membership::{join_room_by_id_helper, leave_all_rooms, leave_room, validate_and_add_event_id};
 pub(super) use message::*;
